+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+13-05-2022 6.6.86
 - Core: Token dates could be saved in user timezone instead of UTC by old framework.
 - Core: Used unnamed parameters in queries for better performance with mysql native driver
   (without PDO::ATTR_EMULATE_PREPARES) and large inserts